requirements.txt
# Django and REST Framework
Django>=4.2,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.5
django-modeltranslation>=0.18.0

# JWT Authentication
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0
cryptography>=41.0.0

# Social Authentication
dj-rest-auth[with_social]>=5.0.0
django-allauth>=0.58.0

# Database
psycopg2-binary>=2.9.9

# Environment Variables
python-decouple>=3.8

# Validation and Utils
phonenumbers>=8.13.0
django-phonenumber-field>=7.2.0

# Email/SMS Verification
django-otp>=1.3.0
boto3>=1.34.0  # AWS SDK for SNS SMS

# API Documentation
drf-spectacular>=0.26.0

# Celery for async tasks
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
flower>=2.0.0

# Production Server
gunicorn>=21.0.0

# Image Processing
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
factory-boy>=3.3.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0

requests